zero_degree/pyproject.toml
[project]
name = "snowgate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ape-vyper>=0.8.10",
    "eth-ape>=0.8.48",
    "fastapi>=0.129.0",
    "groq>=1.0.0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
]
